+++ src/model/user.rs
@@ -9,6 +9,6 @@ pub struct User {
     pub id: u32,
     pub name: String,
     pub score: u8,
-    pub game_id: Option<u32>,
+    pub game_id: Option<u16>,
     pub ws: Option<Addr<UserSocket>>,
 }
